backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-dotenv
python-multipart